@@ -1,13 +1,13 @@
-import React from 'react';
+import React from "react";
 
-import classes from './SidebarToggle.module.css';
+import classes from "./SidebarToggle.module.css";
 
 const sidebarToggle = (props) => (
-    <div className={classes.SidebarToggle} onClick={props.clicked}>
-        <div></div>
-        <div></div>
-        <div></div>
-    </div>
+  <div className={classes.SidebarToggle} onClick={props.clicked}>
+    <div></div>
+    <div></div>
+    <div></div>
+  </div>
 );
 
 export default sidebarToggle;
