@@ -1,12 +1,14 @@
-import React from 'react';
-import classes from './Button.module.css';
+import React from "react";
+import classes from "./Button.module.css";
 
 const button = (props) => (
-    <button 
-        disabled={props.disabled}
-        className={[classes.Button, classes[props.btnType]].join(' ')}    
-        onClick={props.clicked}
-    >{props.children}</button>
+  <button
+    disabled={props.disabled}
+    className={[classes.Button, classes[props.btnType]].join(" ")}
+    onClick={props.clicked}
+  >
+    {props.children}
+  </button>
 );
 
 export default button;
