@@ -1,61 +1,70 @@
-import React from 'react';
+import React from "react";
 
-import classes from './Input.module.scss';
+import classes from "./Input.module.scss";
 
-const input = ( props ) => {
-    let inputElement = null;
-    const inputClasses = [classes.InputElement];
+const input = (props) => {
+  let inputElement = null;
+  const inputClasses = [classes.InputElement];
 
-    if (props.invalid && props.shouldValidate && props.touched) {
-        inputClasses.push(classes.Invalid);
-    }
+  if (props.invalid && props.shouldValidate && props.touched) {
+    inputClasses.push(classes.Invalid);
+  }
 
-    switch ( props.elementType ) {
-        case ( 'input' ):
-            inputElement = <input
-                type={props.types}
-                name={props.names}
-                className={inputClasses.join(' ')}
-                {...props.elementConfig}
-                value={props.value}
-                onChange={props.changed} />;
-            break;
-        case ( 'textarea' ):
-            inputElement = <textarea
-                className={inputClasses.join(' ')}
-                {...props.elementConfig}
-                value={props.value}
-                onChange={props.changed} />;
-            break;
-        case ( 'select' ):
-            inputElement = (
-                <select
-                    className={inputClasses.join(' ')}
-                    value={props.value}
-                    onChange={props.changed}>
-                    {props.elementConfig.options.map(option => (
-                        <option key={option.value} value={option.value}>
-                            {option.displayValue}
-                        </option>
-                    ))}
-                </select>
-            );
-            break;
-        default:
-            inputElement = <input
-                className={inputClasses.join(' ')}
-                {...props.elementConfig}
-                value={props.value}
-                onChange={props.changed} />;
-    }
+  switch (props.elementType) {
+    case "input":
+      inputElement = (
+        <input
+          type={props.types}
+          name={props.names}
+          className={inputClasses.join(" ")}
+          {...props.elementConfig}
+          value={props.value}
+          onChange={props.changed}
+        />
+      );
+      break;
+    case "textarea":
+      inputElement = (
+        <textarea
+          className={inputClasses.join(" ")}
+          {...props.elementConfig}
+          value={props.value}
+          onChange={props.changed}
+        />
+      );
+      break;
+    case "select":
+      inputElement = (
+        <select
+          className={inputClasses.join(" ")}
+          value={props.value}
+          onChange={props.changed}
+        >
+          {props.elementConfig.options.map((option) => (
+            <option key={option.value} value={option.value}>
+              {option.displayValue}
+            </option>
+          ))}
+        </select>
+      );
+      break;
+    default:
+      inputElement = (
+        <input
+          className={inputClasses.join(" ")}
+          {...props.elementConfig}
+          value={props.value}
+          onChange={props.changed}
+        />
+      );
+  }
 
-    return (
-        <div className={classes.Input}>
-            <label className={classes.Label}>{props.label}</label>
-            {inputElement}
-        </div>
-    );
-
+  return (
+    <div className={classes.Input}>
+      <label className={classes.Label}>{props.label}</label>
+      {inputElement}
+    </div>
+  );
 };
 
 export default input;
