@@ -1,10 +1,10 @@
-import React from 'react';
+import React from "react";
 
-import classes from './Modal.module.css';
-import Auxiliary from '../../../hoc/Auxiliary';
-import Backdrop from '../Backdrop/Backdrop';
+import classes from "./Modal.module.css";
+import Auxiliary from "../../../hoc/Auxiliary";
+import Backdrop from "../Backdrop/Backdrop";
 
-const modal = props => {
+const modal = (props) => {
   // shouldComponentUpdate ( nextProps, nextState ) {
   //     return nextProps.show !== this.props.show || nextProps.children !== this.props.children;
   // }
@@ -15,8 +15,8 @@ const modal = props => {
       <div
         className={classes.Modal}
         style={{
-          transform: props.show ? 'translateY(0)' : 'translateY(-100vh)',
-          opacity: props.show ? '1' : '0'
+          transform: props.show ? "translateY(0)" : "translateY(-100vh)",
+          opacity: props.show ? "1" : "0",
         }}
       >
         {props.children}
