@@ -1,8 +1,6 @@
-import React from 'react';
-import classes from './Spinner.module.css';
+import React from "react";
+import classes from "./Spinner.module.css";
 
-const spinner = () => (
-    <div className={classes.Loader}>Loading...</div>
-);
+const spinner = () => <div className={classes.Loader}>Loading...</div>;
 
 export default spinner;
