@@ -1,13 +1,14 @@
-import React from 'react';
-import classes from './Background.module.css';
+import React from "react";
+import classes from "./Background.module.css";
 // import myVid from '../../../assets/videos/myVid.mp4';
-import myImg from '../../../assets/images/background.jpg';
+import myImg from "../../../assets/images/background.jpg";
 
 const background = () => {
-    return (
+  return (
     <div className={classes.BackgroundWrapper}>
-        <img src={myImg} />
-    </div>)
-}
+      <img src={myImg} />
+    </div>
+  );
+};
 
 export default background;
