@@ -1,5 +1,5 @@
-if (process.env.NODE_ENV === 'production') {
-    module.exports = require('./prod.js');
+if (process.env.NODE_ENV === "production") {
+  module.exports = require("./prod.js");
 } else {
-    module.exports = require('./dev.js');
+  module.exports = require("./dev.js");
 }
