@@ -1,5 +1,5 @@
 module.exports = {
-    stripePublishableKey    : process.env.STRIPE_PUBLISHABLE_KEY,
+    stripePublishableKey    :'pk_live_51J8eeGGhmYf089672u6NJwvMPcEER5uaLCQ3eejbU8FNjl6MCe4JKK5DPp5AJF7OpAHFhgK2uhit046XhSnZUqMG00MQjPQoKZ',
     taxRates                : process.env.TAX_RATES,
 
     proxyTarget: "http://localhost:3002" 
