@@ -1,5 +1,5 @@
-const keys = require('./config/keys');
-const {createProxyMiddleware} = require('http-proxy-middleware');
+const keys = require("./config/keys");
+const { createProxyMiddleware } = require("http-proxy-middleware");
 
 module.exports = function (app) {
   app.use(
