@@ -1,34 +1,29 @@
 export {
-    fetchUser,
-    fetchUsers,
-    connect,
-    auth,
-    fbAuth,
-    setAuthRedirectPath,
-    newAddress,
-    newAddressStart
-} from './auth';
-
-
-export {
-    getItems,
-    addToCart,
-    removeItem,
-    subtractQuantity,
-    addQuantity,
-    loadCart,
-    loadShop,
-    orderBy,
-    newItem,
-    newItemStart,
-    getItemById,
-    getItemByType,
-    deleteItem, 
-    updateItem,
-
-} from './shop';
+  fetchUser,
+  fetchUsers,
+  connect,
+  auth,
+  fbAuth,
+  setAuthRedirectPath,
+  newAddress,
+  newAddressStart,
+} from "./auth";
 
 export {
-    fetchOrders,
-    fetchOrdersStart,
-} from './orders'
+  getItems,
+  addToCart,
+  removeItem,
+  subtractQuantity,
+  addQuantity,
+  loadCart,
+  loadShop,
+  orderBy,
+  newItem,
+  newItemStart,
+  getItemById,
+  getItemByType,
+  deleteItem,
+  updateItem,
+} from "./shop";
+
+export { fetchOrders, fetchOrdersStart } from "./orders";
